se prepara para entrega

+++ src/controllers/products.controllers.js
@@ -1,50 +1,35 @@
 import { productsService } from "../services/index.js";
+import CustomError from "../errors/custom_errors.js";
 
-export const addProduct = async (req, res) => {
+export const addProduct = async (req, res, next) => {
   try {
-    const { body } = req;
+    const { body, user: { user } } = req;
+
     const product = await productsService.getProductByFilter({
       code: body.code,
     });
-    if (product) {
-      return res.status(400).json({ status: "error", payload: "Invalid code" });
+
+    if (product || !productsService.checkProductProperties(body)) {
+      CustomError.createProduct(body)
     }
-    if (!productsService.checkProductProperties(body))
-      return res
-        .status(400)
-        .json({ status: "error", payload: "Missed properties" });
+
+    (user.role != "admin") && (body.owner = (user?._id || user?.id))
     const result = await productsService.addProduct(body);
     res.json({ status: "success", payload: result });
   } catch (e) {
-    console.error("Error:", e);
-    res.status(500).send("Server error");
+    next(e)
   }
 };
 
 export const getProducts = async (req, res) => {
   try {
-    const queryFindParameters = {};
-    const limit = parseInt(req.query?.limit ?? 10);
-    const page = parseInt(req.query?.page ?? 1);
-    const title = req.query?.query || null;
-    const optionsPaginate = { limit, page, lean: true };
-    const sortOrder = req.query?.sort?.toLowerCase();
-    const categoryFilter = req.query?.category;
-    const stockFilter = parseInt(req.query?.stock);
-
-    (sortOrder === "asc" && (optionsPaginate.sort = { price: 1 })) ||
-      (sortOrder === "desc" && (optionsPaginate.sort = { price: -1 }));
-    stockFilter && (queryFindParameters.stock = { $gte: stockFilter });
-    categoryFilter && !title && (queryFindParameters.category = categoryFilter);
-    title && ((queryFindParameters.title = title), (optionsPaginate.page = 1));
-
     const result = await productsService.getProducts(
-      queryFindParameters,
-      optionsPaginate
+      req?.queryFindParameters,
+      req?.optionsPaginate
     );
     res.json({ status: "success", payload: result.docs });
   } catch (e) {
-    console.error("Error:", e);
+    req.logger.error("Error: " + e);
     res.status(500).send("Server error");
   }
 };
@@ -54,7 +39,7 @@ export const getProductById = async (req, res) => {
     const result = await productsService.getProductById(req.params?.pid);
     res.json({ status: "success", payload: result });
   } catch (e) {
-    console.error("Error:", e);
+    req.logger.error("Error: " + e);
     if (e.name == "CastError") return res.status(404).send("Not found");
     res.status(500).send("Server error");
   }
@@ -62,27 +47,41 @@ export const getProductById = async (req, res) => {
 
 export const updateProduct = async (req, res) => {
   try {
-    const {body} = req
+    const { body } = req
+    const { user } = req?.user
+
+    if (user.role != "admin") {
+      const product = await productsService.getProductById(req?.params?.pid)
+      if (product?.owner != (user?._id || user?.id) || !product?.owner) return res.status(403).json({ error: "No permission to update this product" })
+    }
+
     if (body._id || body.code)
-      return res.status(400).json({status: "error", payload: "Invalid properties to update"})
+      return res.status(400).json({ status: "error", payload: "Invalid properties to update" })
     const result = await productsService.updateProduct(req.params?.pid, {
       ...body,
     });
-    res.json({status: "success", payload: result})
+    res.json({ status: "success", payload: result })
   } catch (e) {
-    console.error("Error:", e);
+    req.logger.error("Error: " + e);
     if (e.name == "CastError") return res.status(404).send("Not found");
     res.status(500).send("Server error");
   }
 };
 
-export const deleteProduct = async (req,res) => {
+export const deleteProduct = async (req, res) => {
   try {
+    const { user } = req?.user
+
+    if (user.role != "admin") {
+      const product = await productsService.getProductById(req?.params?.pid)
+      if (product?.owner != (user?._id || user?.id)) return res.status(403).json({ error: "No permission to delete this product" })
+    }
+
     const result = await productsService.deleteProduct(req.params?.pid)
-    res.json({status: "success", payload: result})
+    res.json({ status: "success", payload: result })
   }
   catch (e) {
-    console.error("Error:", e);
+    req.logger.error("Error: " + e);
     if (e.name == "CastError") return res.status(404).send("Not found");
     res.status(500).send("Server error");
   }
